crates/sel4-runtime-common: Flatten module structure
