+++ crates/sel4-runtime-common/src/lib.rs
@@ -7,6 +7,9 @@
 #![no_std]
 #![feature(cfg_target_thread_local)]
 
+use sel4_elf_header::{ElfHeader, ProgramHeader};
+use sel4_panicking_env::abort;
+
 mod ctors;
 
 pub use ctors::run_ctors;
@@ -14,22 +17,28 @@ pub use ctors::run_ctors;
 #[cfg(feature = "start")]
 mod start;
 
-#[cfg(any(
-    all(feature = "tls", target_thread_local),
-    all(feature = "unwinding", panic = "unwind")
-))]
-mod phdrs;
+#[cfg(all(feature = "tls", target_thread_local))]
+mod tls;
 
-#[cfg(any(
-    all(feature = "tls", target_thread_local),
-    all(feature = "unwinding", panic = "unwind")
-))]
-pub use phdrs::*;
+#[cfg(all(feature = "tls", target_thread_local))]
+pub use tls::{initialize_tls_on_stack_and_continue, ContArg, ContFn};
 
-#[doc(hidden)]
-pub mod _private {
-    #[cfg(feature = "start")]
-    pub use crate::start::_private as start;
+#[cfg(all(feature = "unwinding", panic = "unwind"))]
+mod unwinding;
+
+#[cfg(all(feature = "unwinding", panic = "unwind"))]
+pub use self::unwinding::set_eh_frame_finder;
+
+pub(crate) fn locate_phdrs() -> &'static [ProgramHeader] {
+    extern "C" {
+        static __ehdr_start: ElfHeader;
+    }
+    unsafe {
+        if !__ehdr_start.check_magic() {
+            abort!("ELF header magic mismatch")
+        }
+        __ehdr_start.locate_phdrs()
+    }
 }
 
 #[cfg(target_arch = "arm")]
@@ -43,4 +52,10 @@ core::arch::global_asm! {
             mrc p15, 0, r0, c13, c0, 2
             bx lr
     "#
+}
+
+#[doc(hidden)]
+pub mod _private {
+    #[cfg(feature = "start")]
+    pub use crate::start::_private as start;
 }
