+++ crates/sel4-runtime-common/src/tls.rs
@@ -14,7 +14,7 @@ use sel4_initialize_tls_on_stack::{
 
 pub use sel4_initialize_tls_on_stack::{ContArg, ContFn};
 
-use crate::phdrs::locate_phdrs;
+use crate::locate_phdrs;
 
 #[allow(clippy::missing_safety_doc)]
 pub unsafe fn initialize_tls_on_stack_and_continue(cont_fn: ContFn, cont_arg: *mut ContArg) -> ! {
